requirements.txt
playwright==1.35.0
pyotp==2.8.0
pandas==2.0.0
reportlab==3.6.12
pyyaml==6.0.1 